+++ src/base/browser/secondary/tree/abstractTree.ts
@@ -1,6 +1,6 @@
-import { ITreeModel, ITreeMouseEvent, ITreeNode } from "src/base/browser/basic/tree/tree";
-import { ITreeListRenderer, TreeItemRenderer } from "src/base/browser/basic/tree/treeListRenderer";
-import { ITreeListWidget, TreeListWidget } from "src/base/browser/basic/tree/treeListWidget";
+import { ITreeModel, ITreeMouseEvent, ITreeNode } from "src/base/browser/secondary/tree/tree";
+import { ITreeListRenderer, TreeItemRenderer } from "src/base/browser/secondary/tree/treeListRenderer";
+import { ITreeListWidget, TreeListWidget } from "src/base/browser/secondary/tree/treeListWidget";
 import { IListItemProvider, TreeListItemProvider } from "src/base/browser/secondary/listView/listItemProvider";
 import { IListTraitEvent } from "src/base/browser/secondary/listWidget/listTrait";
 import { IListMouseEvent } from "src/base/browser/secondary/listWidget/listWidget";
