+++ src/base/browser/secondary/tree/tree.css
