+++ src/base/browser/secondary/tree/treeListRenderer.ts
@@ -1,4 +1,4 @@
-import { ITreeNode } from "src/base/browser/basic/tree/tree";
+import { ITreeNode } from "src/base/browser/secondary/tree/tree";
 import { IListViewMetadata, IListViewRenderer, RendererType } from "src/base/browser/secondary/listView/listRenderer";
 
 /**
