+++ test/base/browser/basic/tree/multiTreeModel.test.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert';
-import { MultiTreeModel } from 'src/base/browser/basic/tree/multiTreeModel';
-import { ITreeNode } from 'src/base/browser/basic/tree/tree';
+import { MultiTreeModel } from 'src/base/browser/secondary/tree/multiTreeModel';
+import { ITreeNode } from 'src/base/browser/secondary/tree/tree';
 import { ISpliceable } from 'src/base/common/range';
 
 function toList<T>(arr: T[]): ISpliceable<T> {
